Put Lambda function code in its own file
